Add files via upload
A Lindermayer System (L-System) is a mathematical grammar system which builds upon itself. 
The L-System consists of an alphabet, a string called the Axiom and a list of production rules
L-Systems are useful for modelling self-similar growth since previous levels of the L-Systenm 
Create more complex versions in the next level, in accordance to its production rules

The list of rules for an L-System can be made as a dictionary in JS 
That takes a character and outputs a string. 

These alphabetical outputs can be interpreted as drawing functions such as forward or rotate
And this creates a beautiful visualisation of the L-System